+++ packages/akashic-cli-serve/src/client/common/ClientContentLocator.ts
@@ -1,0 +1,8 @@
+import { ContentLocator } from "../../common/ContentLocator";
+
+export class ClientContentLocator extends ContentLocator {
+	asAbsoluteUrl(): string {
+		const host = this.host || window.location.origin;
+		return host + this.asRootRelativeUrl();
+	}
+}