+++ packages/akashic-cli-serve/src/client/store/LocalInstanceEntity.ts
@@ -2,6 +2,8 @@ import {action, observable, computed} from "mobx";
 import {Trigger} from "@akashic/trigger";
 import {TimeKeeper} from "../../common/TimeKeeper";
 import {Player} from "../../common/types/Player";
+import {ClientContentLocator} from "../common/ClientContentLocator";
+import * as ApiRequest from "../api/ApiRequest";
 import {GameViewManager} from "../akashic/GameViewManager";
 import {PlayEntity} from "./PlayEntity";
 import {CoePluginEntity, CreateCoeLocalInstanceParameterObject} from "./CoePluginEntity";
@@ -22,10 +24,11 @@ const toAgvExecutionMode = (() => {
 
 export interface LocalInstanceEntityParameterObject {
 	gameViewManager: GameViewManager;
-	contentUrl: string;
+	contentLocator: ClientContentLocator;
 	executionMode: ExecutionMode;
 	play: PlayEntity;
 	player: Player;
+	resizeGameView?: boolean;
 	argument?: any;
 	playToken?: string;
 	playlogServerUrl?: string;
@@ -45,11 +48,12 @@ export class LocalInstanceEntity implements GameInstanceEntity {
 
 	readonly play: PlayEntity;
 	readonly coePlugin: CoePluginEntity;
-	readonly contentUrl: string;
+	readonly contentLocator: ClientContentLocator;
 
 	private _timeKeeper: TimeKeeper;
 	private _gameViewManager: GameViewManager;
 	private _agvGameContent: agv.GameContent;
+	private _resizeGameView: boolean;
 
 	constructor(params: LocalInstanceEntityParameterObject) {
 		this.onStop = new Trigger<LocalInstanceEntity>();
@@ -57,9 +61,11 @@ export class LocalInstanceEntity implements GameInstanceEntity {
 		this.executionMode = params.executionMode;
 		this.play = params.play;
 		this.isPaused = false;
-		this.contentUrl = params.contentUrl;
+		this.contentLocator = params.contentLocator;
 		this._timeKeeper = new TimeKeeper();
 		this._gameViewManager = params.gameViewManager;
+		this._resizeGameView = !!params.resizeGameView;
+
 		const playConfig: agv.PlaylogConfig = {
 			playId: this.play.playId,
 			executionMode: toAgvExecutionMode(this.executionMode),
@@ -74,7 +80,7 @@ export class LocalInstanceEntity implements GameInstanceEntity {
 			playConfig.playToken = params.playToken;
 		}
 		this._agvGameContent = this._gameViewManager.createGameContent({
-			contentUrl: this.contentUrl,
+			contentLocator: this.contentLocator,
 			player: {
 				id: this.player.id,
 				name: this.player.name
@@ -107,6 +113,13 @@ export class LocalInstanceEntity implements GameInstanceEntity {
 	}
 
 	async start(): Promise<void> {
+		if (this._resizeGameView) {
+			const url = this.contentLocator.asAbsoluteUrl();
+			const contentJson = await ApiRequest.get<{ content_url: string }>(url);
+			const gameJson = await ApiRequest.get<{ width: number, height: number }>(contentJson.content_url);
+			this._gameViewManager.setViewSize(gameJson.width, gameJson.height);
+		}
+
 		await this._gameViewManager.startGameContent(this._agvGameContent);
 		this._timeKeeper.start();
 	}