+++ packages/akashic-cli-serve/src/client/store/Store.ts
@@ -2,11 +2,13 @@ import {observable, action} from "mobx";
 import * as queryString from "query-string";
 import {Player} from "../../common/types/Player";
 import {SandboxConfig} from "../../common/types/SandboxConfig";
+import {ClientContentLocator} from "../common/ClientContentLocator";
 import {PlayEntity} from "./PlayEntity";
 import {PlayStore} from "./PlayStore";
 import {LocalInstanceEntity} from "./LocalInstanceEntity";
 import {DevtoolUiStore} from "./DevtoolUiStore";
 import {ToolBarUiStore} from "./ToolBarUiStore";
+import {NotificationUiStore} from "./NotificationUiStore";
 import {storage} from "./storage";
 import {StartupScreenUiStore} from "./StartupScreenUiStore";
 
@@ -14,9 +16,10 @@ export class Store {
 	@observable playStore: PlayStore;
 	@observable toolBarUiStore: ToolBarUiStore;
 	@observable devtoolUiStore: DevtoolUiStore;
+	@observable notificationUiStore: NotificationUiStore;
 	@observable startupScreenUiStore: StartupScreenUiStore;
 	@observable player: Player | null;
-	@observable contentId: number; // 多分storage辺りに置く方がよさそうだが一旦動かすこと優先でここに置いておく
+	@observable contentLocator: ClientContentLocator; // 多分storage辺りに置く方がよさそうだが一旦動かすこと優先でここに置いておく
 
 	@observable currentPlay: PlayEntity | null;
 	@observable currentLocalInstance: LocalInstanceEntity | null;
@@ -25,14 +28,12 @@ export class Store {
 	@observable argumentsTable: { [name: string]: string };
 
 	constructor() {
-		this.contentId = 0;
 		const query = queryString.parse(window.location.search);
-		if (query.id) {
-			this.contentId = parseInt(query.id, 10);
-		}
+		this.contentLocator = new ClientContentLocator({ contentId: (query.id != null) ? query.id : "0" }); // TODO xnv bootstrapから渡す方が自然では？
 		this.playStore = new PlayStore();
 		this.toolBarUiStore = new ToolBarUiStore();
 		this.devtoolUiStore = new DevtoolUiStore();
+		this.notificationUiStore = new NotificationUiStore();
 		this.startupScreenUiStore = new StartupScreenUiStore();
 		this.player = { id: storage.data.playerId, name: storage.data.playerName };
 		this.currentPlay = null;