+++ packages/akashic-cli-serve/src/server/common/ServerContentLocator.ts
@@ -1,0 +1,9 @@
+import { ContentLocator } from "../../common/ContentLocator";
+import { serverGlobalConfig } from "./ServerGlobalConfig";
+
+export class ServerContentLocator extends ContentLocator {
+	asAbsoluteUrl(): string {
+		const host = this.host || `http://${serverGlobalConfig.hostname}:${serverGlobalConfig.port}`;
+		return host + this.asRootRelativeUrl();
+	}
+}